fix(api): cast validated per_page to int

+++ app/Http/Controllers/Api/NotificationController.php
@@ -17,7 +17,7 @@ final class NotificationController
 
     public function index(PaginatedRequest $request, FormatNotificationRow $format): JsonResponse
     {
-        $perPage = $request->validated()['per_page'] ?? 20;
+        $perPage = (int) ($request->validated()['per_page'] ?? 20);
 
         $paginator = $request->user()->notifications()->simplePaginate($perPage);
